pyproject.toml
[tool.poetry]
name = "commercial-network"
version = "0.1.0"
description = ""
authors = ["Tsurkan <mikh.tsurkan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
